Creado kinich y mejorado algo de legibilidad de codigo
He creado al kinich he "balanceado" un poco las stats y visualmente he añadido mas prints para que de momento en consola se vea bien mas adelante

+++ teyvat-showdown/src/game/Flins.java
@@ -27,7 +27,7 @@ public class Flins extends Character {
 		 */
 		
 		int baseDamage = this.getBaseDamage(enemy);
-		int modifiedAttack = (int) ((int) (double) (baseDamage) * 1.8);
+		int modifiedAttack = (int) (baseDamage * 1.8);
 		double critDamageChance = Math.random();
 		
 		if (critDamageChance <= this.critChance) {
